Eked Email Upholding
Eked email upholding, so now an email won't be awended when nitten. When an @ is nitten outside of an email, it will be wharved to be 'taay' (that is, the awending for 'at', since I couldn't afand how to make it work otherwise).

Eked 'unawendFromPigLatin' to Awend.java for email reading and keeping, as well as mightly toward unawending marks.

Eked threads for Setup and Highfasts for a plotted setup thread being eked.

Wharved the nemming given in Awend.java to glass new thread steads for wises.

Stirred some dealt dern into new wises.

I was moned of the do-while ring, so I have wharved the main wise to now brook that instead of a while ring.

Fornoom ynoot 'shouldThroughwon' boolean, now brooks 'shouldWork' alone.